// backend/src/interfaces/tmdb-movie.interface.ts
export interface TMDBMovieResponse {
    page:          number;
    results:       TMDBMovie[];
    total_pages:   number;
    total_results: number;
}

export interface TMDBMovie {
    adult:             boolean;
    backdrop_path:     string;
    id:                number;
    title:             string;
    original_title:    string;
    overview:          string;
    poster_path:       string;
    media_type:        MediaType;
    original_language: OriginalLanguage;
    genre_ids:         number[];
    popularity:        number;
    release_date:      Date;
    softcore:          boolean;
    video:             boolean;
    vote_average:      number;
    vote_count:        number;
}

export enum MediaType {
    Movie = "movie",
}

export enum OriginalLanguage {
    En = "en",
}

// Generated by https://quicktype.io

export interface TMDBMovieDetails {
    adult:                 boolean;
    backdrop_path:         string;
    belongs_to_collection: BelongsToCollection;
    budget:                number;
    genres:                Genre[];
    homepage:              string;
    id:                    number;
    imdb_id:               string;
    origin_country:        OriginCountry[];
    original_language:     OriginalLanguage;
    original_title:        string;
    overview:              string;
    popularity:            number;
    poster_path:           string;
    production_companies:  ProductionCompany[];
    production_countries:  ProductionCountry[];
    release_date:          Date;
    revenue:               number;
    runtime:               number;
    softcore:              boolean;
    spoken_languages:      SpokenLanguage[];
    status:                string;
    tagline:               string;
    title:                 string;
    video:                 boolean;
    vote_average:          number;
    vote_count:            number;
    credits:               Credits;
    similar:               Similar;
    videos:                Videos;
}

export interface BelongsToCollection {
    id:            number;
    name:          string;
    poster_path:   null;
    backdrop_path: string;
}

export interface Credits {
    cast: Cast[];
    crew: Cast[];
}

export interface Cast {
    adult:                boolean;
    gender:               number;
    id:                   number;
    known_for_department: string;
    name:                 string;
    original_name:        string;
    popularity:           number;
    profile_path:         null | string;
    cast_id?:             number;
    character?:           string;
    credit_id:            string;
    order?:               number;
    department?:          string;
    job?:                 string;
}

export interface Genre {
    id:   number;
    name: string;
}

export enum OriginCountry {
    Us = "US",
}

export interface ProductionCompany {
    id:             number;
    logo_path:      string;
    name:           string;
    origin_country: OriginCountry;
}

export interface ProductionCountry {
    iso_3166_1: OriginCountry;
    name:       string;
}

export interface Similar {
    page:          number;
    results:       SimilarResult[];
    total_pages:   number;
    total_results: number;
}

export interface SimilarResult {
    adult:             boolean;
    backdrop_path:     null | string;
    genre_ids:         number[];
    id:                number;
    title:             string;
    original_language: string;
    original_title:    string;
    overview:          string;
    popularity:        number;
    poster_path:       string;
    release_date:      string;
    softcore:          boolean;
    video:             boolean;
    vote_average:      number;
    vote_count:        number;
}

export interface SpokenLanguage {
    english_name: string;
    iso_639_1:    string;
    name:         string;
}

export interface Videos {
    results: VideosResult[];
}

export interface VideosResult {
    iso_639_1:    OriginalLanguage;
    iso_3166_1:   OriginCountry;
    name:         string;
    key:          string;
    site:         Site;
    size:         number;
    type:         Type;
    official:     boolean;
    id:           string;
    published_at: Date;
}

export enum Site {
    YouTube = "YouTube",
}

export enum Type {
    BehindTheScenes = "Behind the Scenes",
    Featurette = "Featurette",
    Teaser = "Teaser",
    Trailer = "Trailer",
}